Move resolveComment to the select package
